Added basics for backend
