+++ .github/workflows/go.yml
@@ -19,7 +19,4 @@ jobs:
         go-version: 1.15
 
     - name: Build
-      run: go build -v ./...
-
-    - name: Test
-      run: go test -v ./...
+      run: go build -v ../../Backend/main.go
